Don't drop events while requesting CPU usage from driver
Every second, we request per-process CPU usage from the driver
which generates a stream of sinsp meta events. Unfortunately,
whenever we do this, we drop the event we have received from scap_next
just before we decided to get the CPU usage data.

This patch moves the logic to a separate method, where an early return
does not prevent handling the received event.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>