Configurable thread table purging
Exited threads are not removed immediately, because we want the thread
to be present in the table for handling e.g. exit() events after
sinsp_parser::process_event() finishes but before the next call
to sinsp::next().

The delayed removal of an exited thread can either happen in sinsp::next()
itself, or can be handled by the sinsp consumer (if the consumer wants
to hold on to the exited threads even longer).

Originally, switching between these two behaviors was switched at build
time, using `#define HAS_ANALYZER`. This patch makes it configurable
at runtime.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>