Partially automate cri.proto handling
musl libc defines stdin, stdout and stderr as macros, so the names
cannot be used as field/method names in generated CRI protobuf
code.

Before this change, we patched cri.proto with a static patch, which
is tedious while upgrading cri.proto.

Introduce a script, `userspace/libsinsp/scripts/update-cri-proto`,
that downloads the latest version, replaces the problematic
names and saves a new version of `userspace/libsinsp/cri.proto`.

Note: this script requires bash and possibly GNU sed.

Note 2: we also replace some occurrences of stdin/out/err in
the comments, which looks weird but is harmless.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>