WIP on plugin API improvements
Some improvements to the plugin api:

 - Note that all char * returned by functions are assumed to be
   allocated and will be free()d. This allows for maximum
   compatibility with plugins written in go/other languages with a C
   interface. For example, go's C.CString returns an allocated charbuf
   and pointers to go memory isn't allowed.
 - Add get_contact() to return contact info for plugin author.
 - Add get_version() to return a version string for the plugin.
 - Add get_event_source() to return a name to describe events returned
   by this source plugin, from calls to next().
 - Add get_extract_event_sources() to return a list of sources that
   this extractor plugin can handle. This is optional and if NULL all
   events will be passed to the extractor plugin.

The previously-unused sinsp_plugin_desc held some of this info, add
the new strings to this object.

Try to be more complete wrt #includes in header files as well as using
std:: rather than assuming a prior "using namespace std".

Also start to isolate direct access to the ss_plugin_info struct. This
is important because the char buffers from struct calls are all
allocated and we don't want to pass along the free() responsibilities
to filterchecks, etc. This isn't complete yet and we need to decide
how much to proxy through sinsp_plugin objects.

Rename get_source_plugin_by_id to just get_plugin_by_id, as it holds
both source and extractor plugins.

This version mostly compiles but haven't tried to run it yet.