Changes to version/info handling
Also save the required api version when loading a plugin. It's also
checked very early on before any plugin object is created, so the
version is already valid and compatible, but saving it might be useful
for debugging.

Change sinsp_plugin::plugin_infos to return structured info instead of
a list, as callers might want to get specific info from the set of
loaded plugins for stuff like compatibility.