Save state/instance ptrs in C structs
Save plugin state/instance structs only in the
{source,extractor}_plugin_info structs. This was important for the
instance struct, as libscap might call open on its own when a plugin
is set as an inspector's input plugin. The state was mostly for
consistency.

This avoids some special handling in sinsp::open to copy the result of
plugin_init back to the plugin object in libscap.