Bug fixes for extractor plugins
If an extractor plugin names no compatible sources, it works for every
source.

Clear the plugin states so a loaded but not-initialized plugin object
doesn't call the destroy function when the sinsp_plugin object is
deleted.

When parsing filtercheck field names, the provided string might be the
whole output string. Only consider the part up to the first space
after the field name.