Update get_fields to pass an "argument required" flag
filterchecks have an EPF_REQURIES_ARGUMENT flag that if set, ensures
that when a field is parsed from a filter string that the argument was
provided.

Plugins didn't have any way to express that property, though.

Update the plugin api to reflect this in the returned json and to set
the flag for a created filtercheck object if set.