Switch to a single extract func w/ multiple fields
Switch the plugin interface to use a single extract function
extract_fields, which also takes a variable-length array of fields to
extract. This will be useful to allow for batch field extracts via a
single plugin method.

A new struct ss_plugin_extract_field holds a field/arg/type, and the
plugin fills in the field_present/{res_str,res_u64}. This is used for
both the blocking and async interfaces.c

At the framework level, this simplifies the plugin object a bit, as it
only needs to handle a single extract function.

Also, remove support for event generation via
next()/next_batch(). This wasn't used anyway--libscap is responsible
for event generation for now.