Also skip EF_OLD_VERSION events from simple consumer
The event table has a few "old" versions of some events, where the
libs needed new params for an event, and did that by defining a new
event. PPM_SYSCALL_EXECVE_* are good examples, where the parameters
kept changing as we wanted to track additional info during execs.

There's no reason to consider these old events with a simple consumer,
so exclude events with this flag.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>