Add a plugin API function to free allocated memory
Add a plugin API function to free any memory allocated by the
plugin. The dynamic symbol is plugin_free_mem(). In create_plugin(),
try to resolve plugin_free_mem first and return an error if it can't
be found. Otherwise, save it in the plugin_info structs.

In other places, where the framework was calling free() call the
resolved symbol instead.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>