Allow for alt. list of filterchecks when creating filter/formatter
When creating filter/formatter factories, or when creating an
individual filter/formatter, allow specifying an alternate list of
available filterchecks.

This will allow segregating fields used by plugins, which only work
with a given plugin event, from fields used by syscalls, which work
with the "default" set of filterchecks in g_filterlist.

The sinsp_filter_check_list object is a good match for this, but it
represents both a list of filter_checks and also has the default set
of filterchecks for syscalls baked in.

With the notion of factories, we need the first part but don't want
the second. So make up a base class filter_check_list that has the
first part. sinsp_filter_check_list derives from the base class and
has the built-in filterchecks.

Also move it to its own file so it's possible to include the
declaration without including filterchecks.h.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>