Don't define strlcpy when already defined (musl libc)
Falco has a build variant that uses musl libc, and musl libc already
defines strlcpy, so we don't want strlcpy when compiling with musl
libc.

So check for it at cmake time and if found set HAVE_STRLCPY. And only
include the one in strlcpy.h if HAVE_STRLCPY is not defined.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>