Add cleaner built-in way to display filterclass field infos
Add methods to filter_fieldclass_info to return a pretty-printed
representation of the field class, as a string. The output looks like
the following:

-------------------------------
Field Class:                  evt (All event types)
Description:                  These fields can be used for all event types
Event Sources:                aws_cloudtrail syscall

evt.num                       (Type: UINT64) event number.
...

Internally, it uses stringstreams to tokenize long strings like the
description, etc, and wrap them as needed. It also cleans up the
output so words aren't broken across lines.

It has options to print verbose information (adding Types for each
field), and if a set of event sources was provided, those are printed
as well.

This will allow the --list outputs of falco, as well as other programs
that use the libs, to have a single representation of fields.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>