Parse field properties + honor "hidden" property
Actually parse the "properties" property of field descriptions that
come back from plugin_get_fields(). The acceptable values are any of
"hidden", "conversation", or "info". Only "hidden" actually has
meaning to this plugin framework, and sets the EPF_TABLE_ONLY flag of
the filtercheck that gets created.

Also, when converting a list of filterchecks to a list of
filter_fieldclass_info structs, if the EPF_TABLE_ONLY flag is set, add
a (string) tag "EPF_TABLE_ONLY".

Follow-on commits will honor this tag and not display the field.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>