update: extend support to PT_CHARBUFARRAY parameter type
The event parameter type PT_CHARBUFARRAY was defined but only used internally.
Now, code it has been changed to be fully supported like the other parameter types.
PT_CHARBUFARRAY represents a list of one or more strings. Although not fully documented,
this was already supported by Falco with the k8s audit log filter checks. Now, this needs
to be supported natively in libsinsp in order to allow plugins to use list filtercheck types
too. Morover, libsinp is also the right place as it bundles the logic for all the other
event parameter types.

Filterchecks of type PT_CHARBUFARRAY only support list-based operators: 'in', 'intersects', and 'pmatch'.
Using it with other operators leads to an error.

If formatted as a string, the output is a list such as "(item1,item2,...,itemN)". This is not safe for
string escaping, and creates an ambiguity on the ',' separator character. However, this is how Falco implemented
this already, and I reported it faithfully for backward compatibility.

If formatted as a json, the output is a json-formatted list.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>