update: add filtercheck EPF_IS_LIST flag to explicitly indicate list types
The EPF_IS_LIST flag explicitly states that a certain filter check is a list type. If set, the type of the filtercheck (e.g. uint64, string, ...)
indicates the types of all the values in the list.Filtercheck without the EPF_IS_LIST flags are expected to always extract 0 or 1 values from events.
Instead, filterchecks with EPF_IS_LIST will always extract a list containing 0 to N values.

List-type filtercheks are only compatible with the list-related
operators 'in' and 'intersects', and an error is thrown if this is not respected. On the other hand, 'in' and 'intersect' remain compatible with filterchecks
without the EPF_IS_LIST flag, as they are right now.

The plugin API is now extended to allow plugins to specify list fields of each supported type (uint64, string) through the 'plugin_get_fields' function.
In the returned json array, plugins can specify the following field types: 'string', uint64', '[]string', 'uint64'. The '[]' prefix is used that a field
is a list of a certain value type.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>