new(libsinsp): create class definitions for filter ASTs
Right now, the notion of filter text parsing is mixed with its evaluation. In fact, what we currently
use as ASTS is the filtercheck hierarchy, which is also the same data structure used for filter evaluation
at runtime. This is confusing, makes the code sticky, and adds unnecessary complexity in the codebase.
This commit introduces an ad-hoc hierarchy, compact but more descriptive, that is only used for filter
textual parting.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>