Detect FIFO files and support seek() via a memory buffer
When opening capture files, stat()/fstat() to determine if the file is
a FIFO (which normally doesn't support seeks).

If the file is a FIFO, borrow the approaches from
https://github.com/falcosecurity/libs/pull/159 to support limited
seek()s via a memory buffer instead of fseek().

This version wipes the memory buffer on every read, so the seeking is
limited to the width of the last read. This isn't very flexible, but
it does happen to allow reading capture files with the limited seek()s
it does.