DO NOT MERGE: temporarily allow zlib for mac builds
There are a bunch of cmake definitions that assume that if
MINIMAL_BUILD is set, then USE_ZLIB/USE_BUNDLED_ZLIB is not set. Not
sure why this change was made, but as long as USE_BUNDLED_DEPS=ON,
sysdig builds just fine with zlib.

Temporarily allow this to allow reading potential remote compressed
captures.