Track container user info and expose via user.name
Changes to support tracking how a container was configured with an
initial user and make that info available as user.name for
CONTAINER_JSON events:

1. Add a "container user" field m_container_user to container_info. By
   default, the value is "<NA>".
2. In the docker and cri container engine resolvers, parse any
   configured user info out of the json response and set
   m_container_user.
3. Serialize the parsed username to the json blob that comprises a
   CONTAINER_JSON event, and parse it out of the json blob when parsing a
   CONTAINER_JSON event.
4. When creating the fake threadinfo that is attached to a container
   event, also set m_exe to "container:<id>".
5. For the proc.name filtercheck, if the event type is container_json,
   return not the thread uid but the container user.

This ends up being more robust in the face of containers where the
initial process might exec and then setuid than a different user. This
tracks the configured user rather than the uids of processes in the
container, which might change.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>