fix(userspace/libsinsp): fix filter comparisons for PT_BYTEBUF fields
Fields of type PT_BYTEBUF (so far, only evt.buffer) require the internal m_val_storage_len
variable to be initialized and passed-in to the comparison functions in order to implement
the supported filtering operators. Due to some recent refactorings, the value was not passed
down anymore and every check on evt.buffer returned true due to memory comparisons being
performed with length 0.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>