fix(driver): drop EF_USES_FD and EF_MODIFIES_STATE flags for uring syscalls
The ENTER events of io_uring_enter and io_uring_register syscalls had the
EF_USES_FD. But since both syscalls return 0 parameters, it causes the the libs
to crash with the following error:
Runtime error: vector::_M_range_check: __n (which is 0) >= this->size() (which is 0). Exiting.
That is because the code tries to access the file descriptor on the event that has the
EF_USES_FD flag.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>