fix(libsinsp): Don't overwrite good container metadata with bad
PR#231 caused a regression with container metadata:

When a container engine reported a failure *after* we have already
received the full metadata from a different engine, the call
to add_container for the failed lookup overwrote the metadata
coming from the good lookup.

Note: this brings the code into agreement with the comment
that says we should follow the same logic as parsing container
JSON events.

Note: based on code inspection, the issue was supposed to happen
only while the inspector isn't fully initialized, but we've observed
it also while the event loop was up and running.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>