fix(k8s-client) handle network related exceptions
It could happen that k8s api-server throttles the initial requests done to retrieve the k8s metadata.
When this happens we do not exit, but catch the exception generated by the temporary error and handle
it.
We reset the state of the socket_handler, meaning that before it is used the connection has to be initialized
again and the same is done for the k8s_handler associated with the socket_handler. The involved k8s_handler is
set in the initial state where it needs to connect to the api-server before using it to retrieve the data. Another solution could
be to destroy and recreate the involved handlers but it would be to much expensive since the throttling could
persist for relatively long periods.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>