feat: Add new return code for filtered events(3/3)
Ref #490

Now that the return code is in place (step 1) and handled in all the
major libsinsp clients (step 2), it's time to actually start returning
the code.

There should be no user-visible effect of this change. Rather, by
providing additional information to the client, the caller should be
able to better react to the behavior of the library.

Signed-off-by: Nathan Baker <nathan.baker@sysdig.com>