Merge upstream pr 688 (#121)
* fix(container_engine): Only return on success or all retries failed

Instead of always returning a result on the first attempt, only return
results on success or when all retries have failed.

This prevents spurious "container" events for incomplete results.

This is especially important when both docker and cri are enabled,
when both must be tried due to the cgroup pattern overlapping, but
only one actually holds the container.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

* Log a warning when empty container infos are returned

When empty container infos are passed up due to all attempts failing,
log a warining. This will help highlight cases when the communication
with the container runtime isn't working properly.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

* Add debug log to note when a lookup is async or sync

The "async_xxx" refers to the code that performs the lookup (we used
to have a separate "docker" engine, but it's been removed.

To make it more clear about whether a lookup is synchronous or
asynchronous, add a debug log.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

* Use bundled valijson for "regular" build

valijson doesn't really have an ubuntu package, so it can't be
preinstalled. Use the bundled valijson instead.

* Add RE2 to container used for builds + tests

This way it will be present when building with
-DUSE_BUNDLED_DEPS=False

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>