chore(scap): Introduce scap_errprintf
Since __attribute__ for compile-time format checking is a gcc extension,
on non-gcc platforms use a hack with a non-evaluated `printf()` call,
hopefully triggering the compiler's `printf()` verification.

The solution was inspired by and based on https://stackoverflow.com/a/73776278
but had to be modified to use a comma expression since we want
`scap_errprintf` to be an expression (e.g. `return scap_errprintf(...)`)

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>