cleanup(scap): remove refresh_proc_table_when_saving from scap
The only thing this flag did was to allow a proc table rescan
in scap_dump_open_gzfile(), when *not* reading from a capture
file (rescanning the proc table when reading a capture doesn't
make sense since it's a no-op).

OSS Sysdig does allow clearing this flag, but only when reading
a capture file, i.e. when it does nothing anyway, so it should be
safe to remove from libscap (the corresponding flag in sinsp
has extra meaning so it stays as is).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>