fix(scap): use strerror_r return value
We're forcing the XSI-compliant version. It's complicated:
https://linux.die.net/man/3/strerror_r

Since the `_POSIX_C_SOURCE` and `_GNU_SOURCE` are fairly
fundamental, play it ultra safe and don't include scap.h with these
two set to a potentially different value from the rest of the code.

This is a good excuse to introduce scap_const.h, which only contains
a few `#define`s used all over the place. I'd expect many of the
`#include "scap.h"` occurrences are for these constants only
so we might be able to speed up rebuilds a  bit.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>