fix: check if the key exists in defer_lookup
async_key_value_source provides a TTL mechanism, the problem is that
when dequeue_next_key is invoked there is no check for the existence
of the key in m_value_map. If the value was pruned due to TTL expiration
it gets put back by std::map::operator[], which triggers the creation of
a new lookup_request.

Signed-off-by: Iacopo Rozzo <iacopo@sysdig.com>