fix(sinsp): support rootless podman containers in systemd services
When rootless podman containers are run as a service,
there's nothing identifying podman in the cgroup as it looks like:
/user.slice/user-<uid>.slice/user@<uid>.service/<unit>/<container_id>
where <unit> is whatever started the container, e.g. foo.service.

Let's hope for the best and assume that all such cgroups are
podman containers.

We can probably narrow the prefix down to ".service/" in the typical
case but as we're already basically guessing, let's keep it generic.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>