Align both {set,get}sockopt to be treated equally wrt UF_NEVER_DROP
Currently, getsockopt is *not* dropped when the driver is in dropping
mode, while setsockopt is dropped (it has the UF_ALWAYS_DROP) flag.

There isn't any reason to treat the two syscalls differently with
respect to dropping mode--they're both infrequent and most
applications are going to do both if they do either.

So remove the UF_ALWAYS_DROP flag for setsockopt so both are included
when in dropping mode, for consistency.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>