Add const_expr_visitor, use it when compiling/cloning/stringing
In some cases, we want to ensure that a visitor does *not* change the
ast. This includes cases where the ast pointer used by the visitor is
read-only.

To support these use cases, add a const_expr_visitor interface where
all the visit() methods take a const argument.

Also add variants of accept() that take const_expr_visitor arguments
and call the const_expr_visitors visit() method.

Compiling, cloning, and stringifying asts are all cases that should
not change the underlying ast, so switch those to use
const_expr_visitor instead of expr_visitor.

A couple of compile private methods had to be changed to take const
arguments. They already didn't modify those arguments, so it was a
safe change.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>