new(sinsp): add raw output to sinsp-example
This is *really* raw (basic header info and a hexdump
of each parameter) but it may come in handy any time we need
to go deeper in analyzing individual events.

Note: to get an actual raw dump, we could use two extra features:
* show the sentinel if available
* verify the total event length vs the lengths of all parameters

Note 2: I really couldn't be bothered to fight with C++ iostreams
to get a usable hex dump so I went for classic C stdio.

To do this, we'd need the access to the raw `scap_evt`, which
would violate the abstraction layering even more.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>