fix(bpf): correct fsconfigX_failure when pushing empty PT_CHARBUF param
The driver test on `s390x` for fsconfigX_failure fails with:

	/root/git/falcosecurity/libs/test/drivers/event_class/event_class.cpp:799: Failure
	Expected equality of these values:
	  size
	    Which is: 1
	  expected_size
	    Which is: 0
	>>>>> length of the param is not correct. Param id = 6

	[  FAILED  ] SyscallExit.fsconfigX_failure (91 ms)

The related filler implementation uses:
````
                /* Parameter 6: value_charbuf (type: PT_CHARBUF) */
                res = bpf_val_to_ring_mem(data, 0, KERNEL);
````
to push an empty param. However the `__bpf_val_to_ring()` implementation
for PT_CHARBUF does not check for `val == 0` and tries to read a
string from address 0.  The probe read str does not fail and returns
a `\0` terminated string cause above failure.

Note that on `s390x`, there is the concept of which low-core (or prefix
area). It is an area of 2 pages at address zero with lots of critical
state information for system operation. Each CPUs has this information
at the same place and it needs to be set up with prefixing / set prefix
instructions.  That means, a zero string is read which results in an
empty/NUL-terminated string.  Note that this behavior might be specific
to s390x and other architectures might not have memory mapped at address
zero.

Note: Trying to use the `bpf_push_empty_param()` in the fsconfig filler
caused an invalid map access BPF verifier.  That's why changing the
filler helper to ignore if val is zero.

Signed-off-by: Hendrik Brueckner <brueckner@de.ibm.com>