driver(bpf,modern_bpf): socketcall remarks on `accept4`
The BPF and modern BPF implementation supports socketcalls based
on the system call identifier.  On s390x, the `accept` system call
does not exist and the implementation maps a `SYS_ACCEPT` to an
`accept4` system call.

The current `accept4` implementation ignores the `flags` parameter
and returns zero.  If this will be changed, there is a need to test
for `SYS_ACCEPT` to avoid reading beyond the passed parameters as
well as to avoid an information leak.

Signed-off-by: Hendrik Brueckner <brueckner@de.ibm.com>