update(cmake): Add support for shared libraries
Add CMake's BUILD_SHARED_LIBS flag to our option list. Build shared
versions of libscap and libsinsp if it's enabled. This should provide
basic support for #762.

Build shared versions of re2 and oneTBB if BUILD_SHARED_LIBS is enabled.
Additional work needs to be done for the rest of our dependencies.

Signed-off-by: Gerald Combs <gerald@wireshark.org>