new(libpman): Improve error logging
When logging an error via libpman, it uses strerror to represent the
error. libbpf in turn uses -ESRCH for certain scenarios when some
objects were not found, e.g. vmlinux or btf id. This makes libpman
report "No such process" together with the error message, which is
confusing.

Add a special case to handle -ESRCH in error reporting logic.

Signed-off-by: Dmitrii Dolgov <9erthalion6@gmail.com>