new(scap): support no_events mode for gVisor
This new flag means that we're *not* going to get any events
from the engine and are using it just for the secondary
effects of scap_open (mostly getting the process table).

This is needed to safely open a new inspector while another
one exists. Otherwise we'd overwrite the gVisor socket
with a new one (which would become inactive the moment we close
the second inspector), breaking all future gVisor connections.

Co-Authored-By: Angelo Puglisi <angelopuglisi86@gmail.com>
Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>