Also save (pointers) to filtercheck values in order
Currently, a filtercheck saves the raw values in m_val_storages and
a pointer + parsed len in m_val_storages_members. Because
m_val_storages_members is an unordered_set, the original order is
lost. This makes it difficult to print out a filter expression and
compare it to the original input string, as the order of checks like
"field in (val1, val2, val3, ...)" are lost.

To help retain this order, add a vector m_vals that saves the pointer
+ length, but in a vector instead of in an ordered_set.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>