Ssprod 23324 add filtercheck testability improvements (#165)
* Add method to retrieve argid from thread filterchecks

Thread filterchecks can have a numeric arg for some fields like
proc.aname/proc.apid. This allows returning an argid for a filtercheck
object. It's -1 if the field doesn't support or doesn't have an arg.

This can be used in unit tests to print/compare filtercheck objects.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

* Also save (pointers) to filtercheck values in order

Currently, a filtercheck saves the raw values in m_val_storages and
a pointer + parsed len in m_val_storages_members. Because
m_val_storages_members is an unordered_set, the original order is
lost. This makes it difficult to print out a filter expression and
compare it to the original input string, as the order of checks like
"field in (val1, val2, val3, ...)" are lost.

To help retain this order, add a vector m_vals that saves the pointer
+ length, but in a vector instead of in an ordered_set.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

* Add a const base_expr_visitor

Following on the changes in
https://github.com/falcosecurity/libs/pull/837, add a const variant of
base_expr_visitor. This allows definining subclasses that want to
traverse an ast read-only without implementng all the methods.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>

---------

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>