cleanup(scap): only touch machine_info in scap_retrieve_machine_info
The two fields (m_driver_procinfo and m_fd_lookup_limit) are not
related to the machine_info in any way. Also, the handle is memset()
to all zeros before calling scap_init_*_int so the explicit
initialization is unnecessary (modulo an exotic arch where NULL is not
represented by all zeros).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>