Add addl(test-only) way to initialize events, setting errorcode
Some tests rely on creating fake events without all of the overhead of
an inspector, etc. To support new tests that rely on
m_errorcode (generally maps to the res field of events), add a new
initializer that passes in a scap + ppm header and errorcode, and just
directly sets m_errorcode.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>