fix(modern_bpf): Fix connect_x prog
The modern probe connect_x behaviour is different from the original ebpf
in how it handles return value. The new one doesn't extract arguments if
retvalue indicates failure, where the original one didn't do such
distinction.

The problem is that it still makes sense to extract arguments in case if
the failure is only EINPROGRESS.

Signed-off-by: Dmitrii Dolgov <9erthalion6@gmail.com>