Add new filterchecks fs.path.* as an umbrella for file operations
Add 6 new filtercheck fields with the prefix fs.path as an umbrella for
various file operations:

- fs.path.name is for file operations that work on a path like open,
  unlink, rmdir, etc.
- for other file operations that have a source and target, like cp,
  symlink, link, mv, etc, there are fields fs.path.source and
  fs.path.target

All of these convert relative paths to absolute paths when needed,
using the thread's cwd.

- fs.path.nameraw/fs.path.sourceraw/fs.path.targetraw are like the above
  but do *not* convert relative paths to absolute.

The fields only work for exit events and only return a value if the
syscall was successful.

The implementation just maps each syscall to a criteria for
success (generally evt.rawarg.res = 0 but can also be evt.rawarg.fd !=
-1 for syscalls that return a fd), and either a filtercheck to return
the path or the source/target.

This makes it easier to write simple rules for a variety of file
operations that have slightly different values for the path and
success.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>