feat(libsinsp): Support reading types for other file descriptors
Currently when trying to perform reverse shell detection in falco we use a rule
which will fire each time stdout/stderr/stdin is replaced with a network
socket, this works, but is less than ideal.

evt.type=dup and fd.num in (0, 1, 2) and fd.type in ("ipv4", "ipv6")

By adding support for looking up the type of a file descriptor other than the
one associated with a syscall event we can write a single rule which will only
fire once all three of stdout/stderr/stdin are sockets.

evt.type=dup and fd.num in (0, 1, 2) and fd.types[0] in ("ipv4", "ipv6") and
fd.types[1] in ("ipv4", "ipv6") and fd.types[2] in ("ipv4", "ipv6")

Signed-off-by: Allan Xavier <allan.xavier@sysdig.com>