cleanup(scap): start pulling out cgroup support to a separate subsystem
Note: we also discover the cgroupfs mountpoint(s) that we don't
currently use, but since we have to walk through the mounts
table to find the cgroup version (and whether there are no v1/v2
mixed scenarios), just leave it in.

Note: in anticipation of cgroup namespace support, we're breaking
out of the container (if running in one) by going through the host's
/proc/1/root. This is a hack but there's no way to request using
the host's cgroup namespace e.g. in Kubernetes so we have to cope.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>